flask/requirements.txt
click==7.1.1
Flask==1.1.2
psycopg2==2.7.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
uWSGI==2.0.18
Werkzeug==1.0.1
